Fix players that got banned from respawning at their death point instead of their spawn point upon unban.
